Changes for startDate > currentDate and Bid  Value changes
Marking Product as is_active as False in case the startDate is greater than currentDate.
Also, for BidIncrement, Disabling user input as number and rather using up and down arrow to increment and decrement the bid Value with respect to Bid Increment Value supplied.
Moreover, changed that the minimum value that user can bet would be equal to the highest Bid till now.